attack/requirements.txt
python-nmap==0.7.1
paramiko==4.0.0